Always coalesce commits if symbols lists are subsets
Commit coalescing is now skipped only if each symbol list has an item
which is not in the other, and --no-symbol-check overrides this. To
speed things up in this sense, the symbols list is now a Set instead of
an Array.

Add some related TODOs and rewrite the warning messages too.