Fix branch handling
The script assumed that when the branches keyword had arguments, these
were in subsequent lines; however, this is not always the case: we can
also have

    branches somebranch;

Fix this by catching any lines _starting_ with 'branches ' rather than
just matching the word. This must be done after skipping over empty
branches line, and by re-processing the line after stripping the initial
'branches ' part so that the :branches action processing only finds the
actual branch.