Transcode non-UTF-8 log messages
git fast-import assumes that the log messages are in UTF-8, and if this
is not the case strange things can happen (such as truncated log message
and whatnot). Allow the user to specify the encoding of log messages,
and in this case transcode all logs from the specified encoding to
UTF-8.