Only store text hashes for non branchpoints
This has the potential to drastically reduce the memory usage for large
files with many revisions.

The text of a commit is no longer needed once it's child commit has been
processed. The memory usage optimization does not work for branches as
these can't be processed reasonably by rcs-fast-export anyways.